Don't depend on Spring AI being on classpath to load tools
